+++ docs_src/reffk/apis/complex_example.py
@@ -9,7 +9,7 @@ models = Registry(database=database)
 
 
 class PostRef(edgy.ModelRef):
-    __model__ = "Post"
+    __related_name__ = "posts_set"
     comment: str
 
     @field_validator("comment", mode="before")