+++ docs_src/reffk/nutshell.py
@@ -8,7 +8,7 @@ models = Registry(database=database)
 
 
 class PostRef(ModelRef):
-    __model__ = "Post"
+    __related_name__ = "posts_set"
     comment: str