+++ edgy/core/db/models/model_reference.py
@@ -1,11 +1,7 @@
-from typing import TYPE_CHECKING, ClassVar, Optional, Type, Union
+from typing import ClassVar
 
 from pydantic import BaseModel
 
-if TYPE_CHECKING:
-    from edgy.core.db.models.model import Model
-
 
 class ModelRef(BaseModel):
-    __model__: ClassVar[Union[Type["Model"], str]]
-    __foreign_key__: ClassVar[Optional[str]] = None
+    __related_name__: ClassVar[str]